\name{binaryDatedTree}
\alias{binaryDatedTree}
\title{Create binary dated tree
binaryDatedTree class, includes heights for each node and other helper variables}
\usage{
  binaryDatedTree(x, ...)
}
\description{
  Create binary dated tree binaryDatedTree class, includes
  heights for each node and other helper variables
}

