\name{calculate.cluster.size.moments.from.tree}
\alias{calculate.cluster.size.moments.from.tree}
\title{Calculate cluster size moments from tree}
\usage{
  calculate.cluster.size.moments.from.tree(bdt, heights)
}
\arguments{
  \item{bdt}{binary dated tree}

  \item{heights}{vector numeric, heights at which to
  calculate cluster sizes}
}
\description{
  Calculate cluster size moments from tree
}

