\name{rcolgem-package}
\alias{rcolgem-package}
\alias{rcolgem}
\docType{package}
\title{
Compute the likelihood of a gene genealogy and simulate genealogies given complex models of a demographic history, such as compartmental models of infectious disease epidemics.
}
\description{
This package provides methods to compute the likelihood of a gene genealogy using structured coalescent models for populations with complex demographic histories. An example of a complex demographic history is the Susceptible-Infected-Recovered model of infectious disease dynamics and variations thereof. 
}
\details{
\tabular{ll}{
Package: \tab rcolgem\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2014-07-22\cr
License: \tab GPL-3 \cr
}
}


\author{
Erik M Volz

Maintainer: Erik M Volz <erik.volz+rcolgem@gmail.com
}
\references{
E. M. Volz, Complex population dynamics and the coalescent under neutrality, Genetics, January, 2012
}

\keyword{ coalescent }

\examples{

# Refer to vignettes for example usage

}
