\docType{data}
\name{SIMULATIONTIMERESOLUTION}
\alias{SIMULATIONTIMERESOLUTION}
\title{this file contains all R functions of the rcolgem package
expects F.(t), G.(t), and Y.(t) to be in namespace}
\format{num 10000}
\usage{
  SIMULATIONTIMERESOLUTION
}
\description{
  this file contains all R functions of the rcolgem package
  expects F.(t), G.(t), and Y.(t) to be in namespace
}
\keyword{datasets}

